+++ src/hooks/useUsers/useUsers.tsx
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from "react";
 import { IUserResponse } from '../../models/IUserResponse';
 import UserService from '../../services/users.service';
 import usePagination from "../usePagination/usePagination";
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
 function useUsers() {
 
@@ -9,13 +11,28 @@ function useUsers() {
     const { totalElements, pageSize, pageNumber, setPageNumber, setTotalElements, setPageSize } = usePagination();
 
     function onDeleteUser(id: number) {
-
-        UserService.deleteUser(id).then((response) => {
-            alert(`coupon ${id} was deleted successfully!`)
-        }).catch((error) => {
-            alert(error.response.data.message);
+        confirmAlert({
+            title: 'Confirm to delete',
+            message: 'Are you sure you want to delete this coupon?',
+            buttons: [
+                {
+                    label: 'Yes',
+                    onClick: () => {
+                        UserService.deleteUser(id).then((response) => {
+                            alert(`coupon ${id} was deleted successfully!`)
+                        }).catch((error) => {
+                            alert(error.response.data.message);
+                        });
+                        window.location.reload();
+                    }
+                },
+                {
+                    label: 'No',
+                }
+            ]
         });
-        window.location.reload();
+
+
     }
 
     useEffect(() => {
